# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Recruitment Manager Api - OpenAPI 3.0.3
  description: |-
    Recruitment Manager Api OpenAPI 3.0.3 specification.
  license:
    name: Proprietary Licensing
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /invitations:
    get:
      summary: List of invitations
      operationId: getAllInvitations
      tags:
        - invitations
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationDto'

    post:
      summary: Send a new invitation
      operationId: sendInvitation
      tags:
        - invitations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationDto'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationDto'
  /invitations/{id}:
    delete:
      summary: Delete invitation by ID
      operationId: deleteInvitation
      tags:
        - invitations
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the invitation
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Invitation deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invitation has been deleted
        '404':
          description: Invitation not found
  /invitations/user-related:
    get:
      summary: Get invitations related with the used
      operationId: getUserRelatedInvitations
      tags:
        - invitations
      parameters:
        - name: userId
          in: query
          required: true
          description: ID of the user who owns the job
          schema:
            type: integer
            format: int64
        - name: email
          in: query
          required: true
          description: email of the user who owns the job
          schema:
            type: string
      responses:
        '200':
          description: List of invitations for candidate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationDto'
  /invitations/recruiter/{recruiterId}:
    get:
      summary: Get invitations by recruiter ID
      operationId: getInvitationsByRecruiterId
      tags:
        - invitations
      parameters:
        - name: recruiterId
          in: path
          required: true
          description: ID of the recruiter
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of invitations for recruiter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationDto'
  /invitations/by-candidate-and-recruiter:
    get:
      summary: Get invitations by candidate ID and recruiter ID
      operationId: getInvitationsByCandidateAndRecruiter
      tags:
        - invitations
      parameters:
        - name: candidateId
          in: query
          required: true
          description: ID of the candidate
          schema:
            type: integer
            format: int64
        - name: recruiterId
          in: query
          required: true
          description: ID of the recruiter
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of matching invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationDto'
  /invitations/{id}/status:
    patch:
      summary: Update invitation status
      operationId: updateInvitationStatus
      tags:
        - invitations
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the invitation
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/InvitationStatus'
      responses:
        '200':
          description: Invitation status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationDto'
        '400':
          description: Invalid status
        '404':
          description: Invitation not found
  /invitations/received-by-recruiter:
    get:
      summary: Get invitations where job's user ID matches a given value
      operationId: getInvitationsByJobUserId
      tags:
        - invitations
      parameters:
        - name: userId
          in: query
          required: true
          description: ID of the user who owns the job
          schema:
            type: integer
            format: int64
        - name: email
          in: query
          required: true
          description: email of the user who owns the job
          schema:
            type: string
      responses:
        '200':
          description: List of matching invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationDto'
        '404':
          description: No invitations found
  /invitations/received-by-recruited:
    get:
      summary: Get invitations where job's user ID matches a given value
      operationId: getInvitationsByEmail
      tags:
        - invitations
      parameters:
        - name: userId
          in: query
          required: true
          description: ID of the user who owns the job
          schema:
            type: integer
            format: int64
        - name: email
          in: query
          required: true
          description: email of the user who owns the job
          schema:
            type: string
      responses:
        '200':
          description: List of matching invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationDto'
        '404':
          description: No invitations found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InvitationStatus:
      type: string
      enum:
        - sent
        - accepted
        - declined
    InvitationDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        recruiter_id:
          type: integer
          format: int64
        candidate_id:
          type: integer
          format: int64
        job_id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/InvitationStatus'
        created_at:
          type: string
          format: date-time
...